refactor: Rename pdf loading file and adapt props names
We will addother urls to this function next commit
