+++ src/utils/data/index.js
@@ -1,3 +1,3 @@
-export * from "./pdf";
+export * from "./uploads";
 export * from "./json";
 export * from "./studyData";
